Moved leftover SuiteSparse related tests to win64 tests project
