+++ src/MGroup.LinearAlgebra/Exceptions/SuiteSparseException.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 
 namespace MGroup.LinearAlgebra.Exceptions
 {
